+++ relay-server/src/processing/trace_metrics/validate.rs
@@ -1,11 +1,10 @@
 use relay_event_schema::protocol::{MetricType, TraceMetric};
 use relay_protocol::Annotated;
 
-use crate::envelope::WithHeader;
 use crate::managed::{Managed, Rejected};
 use crate::processing::Context;
 use crate::processing::trace_metrics::{
-    Error, ExpandedTraceMetrics, Result, SerializedTraceMetrics,
+    Error, ExpandedTraceMetrics, Result, SerializedTraceMetrics, get_calculated_byte_size,
 };
 use crate::services::outcome::DiscardReason;
 use crate::statsd::RelayDistributions;
@@ -44,7 +43,7 @@ pub fn size(metrics: &mut Managed<ExpandedTraceMetrics>, ctx: Context<'_>) {
     metrics.retain(
         |metrics| &mut metrics.metrics,
         |metric, _| {
-            let size = calculate_size(metric);
+            let size = get_calculated_byte_size(metric);
             let is_too_large = size > max_size_bytes;
 
             relay_statsd::metric!(
@@ -59,27 +58,6 @@ pub fn size(metrics: &mut Managed<ExpandedTraceMetrics>, ctx: Context<'_>) {
             }
         },
     );
-}
-
-/// Calculates the byte size for a trace metric.
-///
-/// Similar to [`relay_ourlogs::calculate_size`].
-fn calculate_size(metric: &WithHeader<TraceMetric>) -> usize {
-    let Some(metric) = metric.value() else {
-        // Same logic we use for logs.
-        return 1;
-    };
-
-    let mut size = 0;
-
-    size += metric.name.value().map_or(0, |s| s.len());
-    size += relay_event_normalization::eap::value_size(&metric.value);
-
-    if let Some(attributes) = metric.attributes.value() {
-        size += relay_event_normalization::eap::attributes_size(attributes);
-    }
-
-    size.max(1)
 }
 
 /// Validates the semantic validity of a trace metric.