+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 **Internal**:
 
+- Calculate and track accepted bytes per individual trace metric item via `TraceMetricByte` data category.
 - Use new processor architecture to process standalone profiles. ([#5741](https://github.com/getsentry/relay/pull/5741))
 
 ## 26.3.1